requirements.txt
beautifulsoup4
langcodes
polib
google-cloud-translate==2.0.1
